Fix block/ban user button
Was not validating (so the list item did the shake animation), then
because of the extra DOM elements for the shake animation, the list var
wasn’t evaluating to the parent ul (but the anim div instead). Fix that
reference for completeness, and fix the original issue by skipping
validation for the ban button (don’t need to validate email because
it’s already an email in the approved/pending list, implying that it
already passed validation).