Fix rare case of removing from pending and adding to approved
if access was previously restricted and a user logged in and placed on
pending, then access was opened up and the user logs in again,
authorizer will remove that user from the pending list and add them to
approved. Before this fix, they weren’t properly removed from the
pending list (probably bug introduced when we refactored user lists
into separate options to solve concurrency issues).