Always clear the google token on logout if it exists
Fix for fringe case: hosted domain is set and different than the
logging in google user, so the google token gets set but the user is
denied access. If we don’t clear the login token, then proceeding
google logins with happen automatically (preventing the user from
switching to a different google account).