Add hd (hosted domain) parameter to Google sign-ins
Allows users to restrict Google logins to a single Google Apps hosted
domain (e.g, mycollege.edu). Note that v1 of the google-api-php-client
does not fully support the hd param, so this release only checks it in
the plugin. When we upgrade to v2 (v2 requires PHP >= 5.6, which
excludes ubuntu 12.04 LTS which includes PHP 5.3), it should work fully.