When adding WordPress users to approved list, always use their existing role
We were assigning the Authorizer default role if no role was specified,
but since users can have “no role for this site.” Some people use this
as a security precaution to block capabilities of problem users, but
keep their content (rather than deleting their user account). Honor the
empty role instead of replacing it with the Authorizer default role.