Fix for bug when logging in as a user with a username conflict
Example case: An approved user logs in via CAS for the first time, but
there is already a WordPress user with the same username, but a
different email address. (Note: username is retrieved from
phpCAS::getUser().) Authorizer incorrectly looks up the currently
logging in user’s email address to use as their username. This commit
fixes that lookup, allowing a new WordPress user to be created with the
email address as the username.