Support LDAP URI in hostname field
ldap_connect() allows either a hostname or an LDAP URI as it’s first
parameter, and the second parameter ($port) is ignored if an LDAP URI
is used. Fully support LDAP URIs by appending a port to the URI if it’s
not specified in the URI, but is specified in the separate Authorizer
setting. Props @timkite for tracking down that behavior.