Add authorizer_automatically_approve_user filter
Can inspect LDAP or CAS or Google attributes during login to determine
if this user should be immediately placed on the approved list. Most
useful for sites that deny access to users and use the pending list to
individually approve users. This gives a method to automate some of
that based on attributes returned from the external login service.