Move CAS redirect hook to avoid headers warning
Enabling the setting “Immediately redirect to CAS login form if it’s the only enabled external service and WordPress logins are hidden” would trigger a “Cannot modify header information - headers already sent” PHP warning if output_buffering isn’t enabled on the web server. This commit moves the redirect hook to before output is sent, but after the CAS redirect in the authenticate hook. Props @steven1350 for reporting the bug.