Fix CAS logout issues
when siteurl is http but site is being viewed over https (but CAS server registration is https), make sure to send the https service URL to the logout endpoint. Similarly, make sure to append a trailing slash to the service URL in case the whitelisted service on the CAS server has that component.