Update CAS check to respect 300 codes
Update the CAS response check to perform a valid response if the CAS
server gives a 300 response code; in this case we acknowledge that the
server is healthy and responding, pointing to another location.  It is
the duty of the system admin to verify that this redirect is valid and
correct.