Update authorizer.php
Update user meta before checking user access. Why ? because if you try to authenticate with a user not yet approved, you get the message : 
"You’re not currently allowed to view this site"
Then, if you click on the back button, it try to disconnect you, but in fact, php session is not destroyed, you are still connected and you can not log with another user unless you destroy phpsessid cookie.