Fix for pending users unable to log out of external service
in custom_logout we check authenticated_by usermeta to determine which external service logout routine to fire. Since pending users don’t have a WP_User, this never fired, so if they tried to log in again they would be immediately redirected to the same message because they were still logged in via the external service.